[core] fix implicit integer conversion warnings
This commit enables the `-Wimplicit-int-conversion` compiler flag for
`ftd`, `mtd`, and `radio` to improve code quality by detecting
potential data loss from implicit type conversions. This is enabled
when clang toolchain is used.

All resulting warnings have been addressed by either:
- Changing variable, parameter, or return types to ensure consistency
  and prevent overflows.
- Adding explicit `static_cast` where the type conversion is intended
  and safe.

+++ src/core/border_router/dhcp6_pd_client.cpp
@@ -696,7 +696,7 @@ void Dhcp6PdClient::SaveServerDuidAndAddress(const Message &aMessage)
     Ip6::Address serverAddress;
 
     SuccessOrAssert(ServerIdOption::ReadDuid(aMessage, serverDuidOffsetRange));
-    mServerDuid.SetLength(serverDuidOffsetRange.GetLength());
+    mServerDuid.SetLength(static_cast<uint8_t>(serverDuidOffsetRange.GetLength()));
     aMessage.ReadBytes(serverDuidOffsetRange, mServerDuid.GetArrayBuffer());
 
     ProcessServerUnicastOption(aMessage, serverAddress);
